Better form submittion POST replies and started comment display code
The form submition post requests now return JSON files with a "valid" status and
a message.  If the status is set to true, the JavaScript replaces the submission
form with a success message.  If it's false, an error message is displayed in an
alert.

Added sanitization for the URL in the form submission POST request.

Began implementing the displayComments app.  At this point, the app gets all the
data from the table as arrays representing rows until there are no rows left. It
then var_dumps the row arrays.

In the JavaScript, began implementing the second half of the display function,
which will display the comments.  Currently creates the comments div and makes
the POST request which will eventually return the code for each comment's div.
Also currently appends the var_dumped data to the comments div.