Specified id for script tag so comments.php can be easily referenced
Specified an id for comments.js's script tag so that it can be gotten via getElementByID().
Created a path variable to the directory in which comments.js is located by getting it by
id and getting its src attribute, then removing "comments.js".  Prepended the path variable
to all references to comments.php.