Created simple default CSS style to make comments more appealing
This style is encapsulated in a function in the JavaScript file, and is called at
the top of the file to make removing all style as easy as commenting or deleting
one line.  The style is inserted before the first child of the head, so any style
tags or external style sheets placed in head will override the default style.